refactor(shared): split product paths from server URLs
