+++ web/src/components/table/peek/peek-observation-detail.tsx
@@ -9,7 +9,7 @@ import {
   traceDetailTitle,
 } from "@/src/features/traces";
 import { resolvePeekTraceParams } from "@/src/components/table/peek/resolvePeekTraceParams";
-import { buildTraceDetailPath } from "@langfuse/shared";
+import { buildTracePath } from "@langfuse/shared";
 import { useRouter } from "next/router";
 import { useRef } from "react";
 
@@ -49,7 +49,7 @@ export const TablePeekViewObservationDetail = (
         traceId: trace.data.id,
         projectId: trace.data.projectId,
         isPublic: trace.data.public,
-        shareUrl: buildTraceDetailPath({
+        shareUrl: buildTracePath({
           projectId: trace.data.projectId,
           traceId: trace.data.id,
           observationId:
