+++ web/src/components/table/peek/peek-trace-detail.tsx
@@ -11,7 +11,7 @@ import {
   shouldClosePeekAfterDelete,
 } from "@/src/components/table/peek";
 import { resolvePeekTraceParams } from "@/src/components/table/peek/resolvePeekTraceParams";
-import { buildTraceDetailPath } from "@langfuse/shared";
+import { buildTracePath } from "@langfuse/shared";
 
 export const TablePeekViewTraceDetail = (
   props: Omit<
@@ -48,7 +48,7 @@ export const TablePeekViewTraceDetail = (
         traceId: trace.data.id,
         projectId: trace.data.projectId,
         isPublic: trace.data.public,
-        shareUrl: buildTraceDetailPath({
+        shareUrl: buildTracePath({
           projectId: trace.data.projectId,
           traceId: trace.data.id,
           timestamp,
